Call out staff flag earlier
With af3d259 I moved the staff stuff together, but upon further
investigation, this was wrong. If you don't set that flag prior to
installing pods you'll get the wrong fonts and be confused. This
happened to me and I fixed it by accident. Later, I was reading the
Podfile and figured out what happened.

This commit calls attention to the flag being set before the pods are
installed.